Move simple protocol type to common
