+++ magma/testing/__init__.py
@@ -1,1 +1,1 @@
-from .utils import check_files_equal, magma_debug_section
+from .utils import check_files_equal, magma_debug_section, SimpleMagmaProtocol
